+++ backend/src/services/sessionExec.js
@@ -5,6 +5,8 @@ const JWT = require('jsonwebtoken')
 const { ForbiddenError, UserInputError } = require('apollo-server-express')
 const { v5: uuidv5 } = require('uuid')
 const mongoose = require('mongoose')
+const dayjs = require('dayjs')
+const { sortBy } = require('ramda')
 
 const { ObjectId } = mongoose.Types
 
@@ -22,6 +24,7 @@ const {
   FEEDBACK_RESOLVED,
   FEEDBACK_RESPONSE_ADDED,
   FEEDBACK_RESPONSE_DELETED,
+  CONFUSION_ADDED,
 } = require('../resolvers/subscriptions')
 const { AUTH_COOKIE_SETTINGS } = require('./accounts')
 
@@ -343,7 +346,6 @@ const deleteFeedback = async ({ sessionId, feedbackId, userId }) => {
  */
 const addConfusionTS = async ({ sessionId, difficulty, speed }) => {
   // TODO: participant auth
-
   const session = await getRunningSession(sessionId)
 
   // if the confusion barometer is not activated, do not allow new additions
@@ -354,18 +356,47 @@ const addConfusionTS = async ({ sessionId, difficulty, speed }) => {
   // push a new timestep into the array
   session.confusionTS.push({ difficulty, speed })
 
-  // save the updated session
-  await session.save()
+  const filteredConfusion = session.confusionTS.filter(
+    (element) => dayjs().diff(dayjs(element.createdAt), 'minute') <= 10
+  )
+  let speedRunning = 0
+  let difficultyRunning = 0
 
-  // extract the saved confusion timestep and convert it to a plain object
+  speedRunning =
+    filteredConfusion.reduce((previousValue, currentValue) => previousValue + (currentValue.speed + 1) * 0.5, 0) /
+    filteredConfusion.length
+  difficultyRunning =
+    filteredConfusion.reduce((previousValue, currentValue) => previousValue + (currentValue.difficulty + 1) * 0.5, 0) /
+    filteredConfusion.length
+
+  if (Number.isNaN(speedRunning)) {
+    speedRunning = 0.5
+  }
+  if (Number.isNaN(difficultyRunning)) {
+    difficultyRunning = 0.5
+  }
+  // overwrite confusionTS data sent to user by filtered and aggregated values
+  session.confusionValues = { speed: speedRunning, difficulty: difficultyRunning }
+
+  // readd mongoDB id-field
+  session.id = session._id
+  session.blocks.forEach((block) => {
+    // eslint-disable-next-line no-param-reassign
+    block.id = block._id
+  })
+
+  // add the computed confusion values to the subscribtion content and
   // then readd the mongo _id field under the id key and publish the result
   // this is needed as redis swallows the _id field and the client could break!
-  // const savedConfusion = session.confusionTS[session.confusionTS.length - 1].toObject()
-  // pubsub.publish(CONFUSION_ADDED, {
-  //   [CONFUSION_ADDED]: { ...savedConfusion, id: savedConfusion._id },
-  //   sessionId,
-  // })
+  const savedConfusion = session.confusionTS[session.confusionTS.length - 1].toObject()
 
+  pubsub.publish(CONFUSION_ADDED, {
+    [CONFUSION_ADDED]: { speed: speedRunning, difficulty: difficultyRunning, id: savedConfusion._id },
+    sessionId,
+  })
+
+  // save the updated session
+  await session.save()
   // return the updated session
   return session
 }
@@ -713,7 +744,12 @@ async function joinQA({ shortname, auth }) {
     verifyParticipantAuthentication({ auth, authenticationMode: settings.authenticationMode, id, participants })
   }
 
-  return settings.isFeedbackChannelActive ? feedbacks.filter((feedback) => feedback.published) : []
+  return settings.isFeedbackChannelActive
+    ? sortBy(
+        (o) => (o.resolvedAt ? -dayjs(o.resolvedAt).unix() : -dayjs(o.createdAt).unix()),
+        feedbacks.filter((feedback) => feedback.published)
+      )
+    : []
 }
 
 /**
@@ -792,6 +828,46 @@ async function resetQuestionBlock({ sessionId, blockId }) {
   return getRunningSession(sessionId)
 }
 
+/**
+ * Aggregates confusion data into what is really required by the client
+ */
+const fetchRunningSessionData = async (userId) => {
+  const user = await UserModel.findById(userId).populate('runningSession')
+
+  const { runningSession } = user
+
+  const filteredConfusion = runningSession.confusionTS.filter(
+    (element) => dayjs().diff(dayjs(element.createdAt), 'minute') <= 10
+  )
+
+  let speedRunning = 0
+  let difficultyRunning = 0
+
+  speedRunning =
+    filteredConfusion.reduce((previousValue, currentValue) => previousValue + (currentValue.speed + 1) * 0.5, 0) /
+    filteredConfusion.length
+  difficultyRunning =
+    filteredConfusion.reduce((previousValue, currentValue) => previousValue + (currentValue.difficulty + 1) * 0.5, 0) /
+    filteredConfusion.length
+
+  if (Number.isNaN(speedRunning)) {
+    speedRunning = 0
+  }
+  if (Number.isNaN(difficultyRunning)) {
+    difficultyRunning = 0
+  }
+  // overwrite confusionTS data sent to user by filtered and aggregated values
+  runningSession.confusionValues = { speed: speedRunning, difficulty: difficultyRunning }
+
+  // readd mongoDB id-field and empty confusionTS
+  runningSession.id = runningSession._id
+  runningSession.blocks = runningSession.blocks.map((block) => ({
+    ...block,
+    id: block._id,
+  }))
+  return runningSession
+}
+
 module.exports = {
   getRunningSession,
   addResponse,
@@ -810,4 +886,5 @@ module.exports = {
   publishFeedback,
   upvoteFeedback,
   reactToFeedbackResponse,
+  fetchRunningSessionData,
 }